Fix Migration

+++ db/migrate/20150908081156_add_value_to_trades.rb
@@ -1,5 +1,5 @@
 class AddValueToTrades < ActiveRecord::Migration
   def change
-    add_column :trades, :value, :double
+    add_column :trades, :value, :decimal
   end
 end
